[RUST] Refactors Tauri's State and adds basic logging
The entire state is now wrapped inside an Arc-Mutex combo, as opposed as
before where only some field where protected.

Adds some logging messages on src/main.rs, using the env_logger crate.
For now, until we find a way to tell tauri how to launch cargo (with
which env variables), in order to see the messages we have to create a
`.env` file an write `RUST_LOG=app` inside.