[RUST] Fixes HTTP Packets collection and adds Content-Encoding support
Adds Content-Encoding handling, unpacking Requests/Responses in case
of gzip, zlib and deflate, and leaving encoded in all other cases

Fixes the way how a packet is determined "complete": I supposed that the
library httparse handled it, but in fact just the Headers are parsed
correctly. As a consequence, also Transer-Encoding is supported.