[TS] [RUST] On-Demand Packet sending and Filtering
The packets are no longer sent to frontend at the exact same time they
are sniffed, but are served on-demand by calling the get_packets Tauri
command (in fact, implementing a server-side pagination).

This implies storing all the packets inside some sort of data structure
(for now, a simple Vec seems enough) which can lead to a huge quantity
of memory used just for that (for now, 1M packets => 2.5 GB). To reduce
the amount of data stored, the payload has been removed from all packets
but link layer ones.

As a consequence, also the filtering process is moved to the backend,
since now all the packets are here. To improve the overall performance
a series of Indexes is used, in form of Vec for Type Filters and in form
of BTreeMap for Strong Filters.

Co-authored-by: Christian Cagnazzo <christiancagnazzo@hotmail.com>