[RUST] Partially fixes the Start-Stop problem
The race condition generated by the Start-Stop sequence is difficult to
remove, especially if we want to reuse the same sniffing Thread.
For now, as a simple workaround, a new Thread is created everytime the
sniffing process is started.
Like that we can have different threads all sniffing from the same
interface in a given time window, but just one that actually sends the
result to the frontend.

The problem that introduces this fix is the fact that if we sniff on a
"empty inyterface", a new thread is created at each Start, and will
never be released, in fact leaking memory.