[RUST] Partial fixes Start-Stop bug and fixes page buttons lagging
- Fixes the way the structure are cleaned when a sniffing process is
stopped. The indexes were cleaned up, but the packets list no.
- Fixed page buttons lagging
- Adds github worflow for release
- **Partially fixes the Start-Stop problem**
The race condition generated by the Start-Stop sequence is difficult to
remove, especially if we want to reuse the same sniffing Thread.
For now, as a simple workaround, a new Thread is created everytime the
sniffing process is started.
Like that we can have different threads all sniffing from the same
interface in a given time window, but just one that actually sends the
result to the frontend.
The problem that introduces this fix is the fact that if we sniff on a
"empty inyterface", a new thread is created at each Start, and will
never be released, in fact leaking memory.

Co-authored-by: Christian Cagnazzo <christiancagnazzo@hotmail.com>